added dependency to pom file and moved testing file to the testing area
