+++ src/test/java/com/revature/AuthServiceTest.java
@@ -1,4 +1,4 @@
-package com.revature.Service;
+package com.revature;
 
 import static org.junit.Assert.assertTrue;
